sql-Agent/requirements.txt
langgraph
pydantic
google-generativeai
google-api-core
streamlit
pandas
loguru
numpy